requirements.txt
torch~=1.9.0
tqdm~=4.62.2
numpy~=1.21.2
Pillow~=8.3.2
PyYAML~=5.4.1
scikit-learn~=0.24.2
scipy~=1.7.1
torch~=1.9.0
tqdm~=4.62.2
opencv-python~=4.5.3.56
matplotlib~=3.4.3
tensorboard~=2.6.0
